+++ lunchmenu/sql/__init__.py
@@ -1,2 +1,5 @@
-from sql import Database
+from sql import Room
 from sql import User
+from sql import Restaurant
+from sql import Database
+